+++ src/ui/elements.ts
@@ -12,6 +12,7 @@ export function getElements() {
     oauthAvatar: required<HTMLSpanElement>("oauth-avatar"),
     oauthUsername: required<HTMLSpanElement>("oauth-username"),
     oauthSignoutBtn: required<HTMLButtonElement>("oauth-signout-btn"),
+    dandisetMessage: required<HTMLParagraphElement>("dandiset-message"),
     dandisetSingle: required<HTMLParagraphElement>("dandiset-single"),
     dandisetSingleText: required<HTMLSpanElement>("dandiset-single-text"),
     dandisetSingleLink: required<HTMLAnchorElement>("dandiset-single-link"),